LayoutTests/imported/w3c/web-platform-tests/css/css-text/text-autospace/text-autospace-ligature-001-expected.txt

国fi
国fi
fi国
fi国
国fi国
国fi国

FAIL test1.assert(0, 1, spacing) assert_approx_equals: expected 46 +/- 0.001 but got 41
PASS test1.assert(1, 3, 0)
FAIL test2.assert(0, 2, spacing) assert_approx_equals: expected 29 +/- 0.001 but got 24
PASS test2.assert(2, 3, 0)
FAIL test3.assert(0, 1, spacing) assert_approx_equals: expected 46 +/- 0.001 but got 41
FAIL test3.assert(1, 3, spacing) assert_approx_equals: expected 29 +/- 0.001 but got 24
PASS test3.assert(3, 4, 0)
